requirements.txt
scipy==1.1.0
torch==0.4.1
numpy==1.15.3
tqdm==4.19.6
matplotlib==2.2.2
Pillow==5.3.0
psutil==5.4.8
PyYAML==3.13
